더블로 바꿈

+++ program/1S3B/Assets/2.Scripts/Manager/TileManager.cs
@@ -17,8 +17,8 @@ public struct SaveTileData
 public struct SaveCropData
 {
     public int cropID;
-    public decimal currentStage;
-    public decimal growRatio;
+    public double currentStage;
+    public double growRatio;
     public int harvest;
     public int deathTimer;
 }
@@ -35,8 +35,8 @@ public class CropData
     public SpriteRenderer cropRenderer;
     public int cropID;
     public Crop plantCrop;
-    public decimal currentStage;
-    public decimal growRatio;
+    public double currentStage;
+    public double growRatio;
     public int harvest;
     public int deathTimer;
 
@@ -45,7 +45,7 @@ public class CropData
     {
         plantCrop = cropDatabase.GetItemByKey(id);
         currentStage = 0;
-        growRatio = plantCrop.AllGrowthStage / (decimal)plantCrop.GrowthTime;
+        growRatio = plantCrop.AllGrowthStage / (double)plantCrop.GrowthTime;
         cropObj = go;
         cropRenderer = cropObj.GetComponentInChildren<SpriteRenderer>();
         cropRenderer.sprite = plantCrop.SpriteList[0];
@@ -286,7 +286,7 @@ public class TileManager : Manager
             else if (groundData[cell].isWater == true)
             {
                 if (tempPlantData.harvest > 0)
-                    tempPlantData.growRatio = (tempPlantData.plantCrop.AllGrowthStage - tempPlantData.plantCrop.StageAfterHarvest) / (decimal)tempPlantData.plantCrop.ReGrowthTime;
+                    tempPlantData.growRatio = (tempPlantData.plantCrop.AllGrowthStage - tempPlantData.plantCrop.StageAfterHarvest) / (float)tempPlantData.plantCrop.ReGrowthTime;
 
                 tempPlantData.currentStage += tempPlantData.growRatio;
